Improve internal api server to support multiple lmcache server (#1387)
* Support config multiple api servers for each included server


* Rebase the thread api


* Add comment


---------
